backend/requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
pillow==11.0.0
python-barcode==0.15.1
pyzbar==0.1.9
qrcode==8.0
sqlparse==0.5.2
tzdata==2024.2
